For Encryption package, add reference to ContractEnforcement file in CosmosTest
