+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/ContractEnforcement.cs
@@ -6,33 +6,12 @@
     using System.IO;
     using System.Linq;
     using System.Reflection;
-    using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Newtonsoft.Json;
     using Newtonsoft.Json.Converters;
 
-    [TestCategory("Windows")]
-    [TestClass]
     public class ContractEnforcement
     {
-        private const string BaselinePath = "DotNetSDKAPI.json";
-        private const string BreakingChangesPath = "DotNetSDKAPIChanges.json";
-
         private static readonly InvariantComparer invariantComparer = new InvariantComparer();
-
-        [TestMethod]
-        public void ContractChanges()
-        {
-            Assert.IsFalse(
-                ContractEnforcement.DoesContractContainBreakingChanges("Microsoft.Azure.Cosmos.Client", BaselinePath, BreakingChangesPath),
-                $@"Public API has changed. If this is expected, then refresh {BaselinePath} with {Environment.NewLine} Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/testbaseline.cmd /update after this test is run locally. To see the differences run testbaselines.cmd /diff"
-            );
-        }
-
-        [TestMethod]
-        public void UniqueKeyUnsealed()
-        {
-            Assert.IsFalse(typeof(UniqueKey).IsSealed);
-        }
 
         private static Assembly GetAssemblyLocally(string name)
         {
@@ -126,7 +105,7 @@
             return root;
         }
 
-        private static bool DoesContractContainBreakingChanges(string dllName, string baselinePath, string breakingChangesPath)
+        public static bool DoesContractContainBreakingChanges(string dllName, string baselinePath, string breakingChangesPath)
         {
             TypeTree locally = new TypeTree(typeof(object));
             ContractEnforcement.BuildTypeTree(locally, ContractEnforcement.GetAssemblyLocally(dllName).GetExportedTypes());
